Aggiunta funzione in VeicoloRepository e cambio definizione di prezzoBase da BigDecimal a Double

+++ backend/src/main/java/it/autoflow/vehicle/entity/Veicolo.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter @Setter
@@ -31,7 +30,7 @@ public class Veicolo {
     private String vin;  // telaio
 
     @Column(nullable = false)
-    private BigDecimal prezzoBase;
+    private Double prezzoBase;
 
     private Integer chilometraggio;      // per usato
